requirements.txt
django
djangorestframework
django-phone-login